+++ src/Modules/Core/DBConstants/Store/StoreSettings.php
@@ -1,0 +1,13 @@
+<?php
+
+// table StoreSettings
+
+namespace Foodsharing\Modules\Core\DBConstants\Store;
+
+class StoreSettings
+{
+	public const USE_PICKUP_RULE_YES = 1;
+	public const USE_PICKUP_RULE_NO = 0;
+	public const PRESS_YES = 1;
+	public const PRESS_NO = 0;
+}