+++ src/Utility/PageHelper.php
@@ -247,7 +247,6 @@ final class PageHelper
 				$group['isAdmin'] = $this->session->isAdminFor($groupId);
 				$group['mayAccessReportGroupReports'] = $this->reportPermissions->mayAccessReportGroupReports($groupId);
 				$group['mayAccessArbitrationGroupReports'] = $this->reportPermissions->mayAccessArbitrationReports($groupId);
-				$group['maySetRegionOptions'] = $this->regionPermissions->maySetRegionOptions($groupId);
 				$group['maySetRegionPin'] = $this->regionPermissions->maySetRegionPin($groupId);
 				$regions[] = $group;
 			} else {